Checkpoint in migration
