+++ components/provider/src/structs/dates.rs
@@ -3,6 +3,7 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/master/LICENSE ).
 
 // Date
+
 use crate::prelude::*;
 
 pub mod key {
@@ -12,14 +13,17 @@ pub mod key {
 
 /// Gets a locale-invariant default struct given a data key in this module's category.
 #[cfg(feature = "invariant")]
-pub(crate) fn get_invariant(data_key: &DataKey) -> Option<DataResponse<'static>> {
-    use crate::invariant::make_inv_response;
+pub fn get_invariant<'d>(
+    data_key: &DataKey,
+    receiver: &mut dyn DataReceiver<'d, 'static>,
+) -> Result<(), DataError> {
     match *data_key {
-        key::GREGORY_V1 => make_inv_response::<gregory::DatesV1>(),
-        _ => None,
+        key::GREGORY_V1 => receiver.receive_invariant::<gregory::DatesV1>(),
+        _ => Err(DataError::UnsupportedDataKey(*data_key)),
     }
 }
 
+/// Gets a boxed DataReceiver capable of receiving a data key in this module's category.
 pub fn get_receiver<'d>(data_key: &DataKey) -> Option<Box<dyn DataReceiver<'d, 'static> + 'd>> {
     match *data_key {
         key::GREGORY_V1 => Some(DataReceiverForType::<gregory::DatesV1>::new_boxed()),
