+++ components/provider/src/structs/mod.rs
@@ -12,16 +12,23 @@ use crate::prelude::*;
 /// For example, if the data key is `plurals/cardinal@1`, a Response with an object of type
 /// `PluralRuleStringsV1` will be returned.
 #[cfg(feature = "invariant")]
-pub(crate) fn get_invariant(data_key: &DataKey) -> Option<DataResponse<'static>> {
-    None //
-        .or_else(|| decimal::get_invariant(data_key)) //
-        .or_else(|| plurals::get_invariant(data_key)) //
-        .or_else(|| dates::get_invariant(data_key)) //
+pub fn get_invariant<'d>(
+    data_key: &DataKey,
+    receiver: &mut dyn DataReceiver<'d, 'static>,
+) -> Result<(), DataError> {
+    Err(DataError::UnsupportedDataKey(*data_key)) //
+        .or_else(|_| decimal::get_invariant(data_key, receiver)) //
+        .or_else(|_| plurals::get_invariant(data_key, receiver)) //
+        .or_else(|_| dates::get_invariant(data_key, receiver)) //
 }
 
-pub fn get_receiver<'d>(data_key: &DataKey) -> Option<Box<dyn DataReceiver<'d, 'static> + 'd>> {
+/// Gets a boxed DataReceiver capable of receiving any known data struct.
+pub fn get_receiver<'d>(
+    data_key: &DataKey,
+) -> Result<Box<dyn DataReceiver<'d, 'static> + 'd>, DataError> {
     None //
         .or_else(|| decimal::get_receiver(data_key)) //
         .or_else(|| plurals::get_receiver(data_key)) //
         .or_else(|| dates::get_receiver(data_key)) //
+        .ok_or_else(|| DataError::UnsupportedDataKey(*data_key)) //
 }
